\name{Cedric IO Utilities}
\alias{readGenome}
\title{Cedric IO Utilities}
\description{
  This utility function provides basic IO by Cedric.
}
\usage{
  readGenome(fn.genome, ex.sh.aa = 0, rm.first.aa = 0)
}
\arguments{
  \item{fn.genome}{under development}
  \item{ex.sh.aa}{under development}
  \item{rm.first.aa}{under development}
}
\details{
  under development
}
\value{
  under development
}
\references{
  \url{https://github.com/snoweye/cubfits/}
}
\author{
  Cedric Landerer \email{cedric.landerer@gmail.com}.
}
\seealso{
  under development
}
\examples{
\dontrun{
suppressMessages(library(cubfits, quietly = TRUE))

}
}
\keyword{cedric}
