\name{Cedric Internal Functions}
\alias{plotBMatrixPosterior}
\alias{plotCUB.NSE}
\alias{plotbin.NSE}
\alias{prop.model.nse}

\title{All Cedric Internal Functions}
\description{All Cedric internal functions}
\references{
  \url{https://github.com/clandere/cubfits/}
}
\author{
  Cedric Landerer \email{cedric.landerer@gmail.com}.
}
\keyword{internal}

